chore: cleanup React package structure
