+++ packages/react-dom-interactions/src/index.ts
@@ -1,107 +1,6 @@
-import * as React from 'react';
-import {
-  useFloating as usePositionalFloating,
-  ComputePositionReturn,
-  VirtualElement,
-} from '@floating-ui/react-dom';
-import useLayoutEffect from 'use-isomorphic-layout-effect';
-import type {
-  FloatingContext,
-  Middleware,
-  Placement,
-  Strategy,
-  ContextData,
-} from './types';
-import {createPubSub} from './createPubSub';
-import {useFloatingTree} from './FloatingTree';
-
-interface CleanupFn {
-  (): void;
-  $$immediate?: boolean;
-}
-
-type Data = Omit<ComputePositionReturn, 'x' | 'y'> & {
-  x: number | null;
-  y: number | null;
-};
-
-export type ReferenceType = Element | VirtualElement;
-
-export type UseFloatingReturn<RT extends ReferenceType = ReferenceType> =
-  Data & {
-    update: () => void;
-    reference: (node: RT | null) => void;
-    floating: (node: HTMLElement | null) => void;
-    refs: {
-      reference: React.MutableRefObject<RT | null>;
-      floating: React.MutableRefObject<HTMLElement | null>;
-    };
-  };
-
-export interface Props<RT extends ReferenceType = ReferenceType> {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-  placement: Placement;
-  middleware: Array<Middleware>;
-  strategy: Strategy;
-  nodeId: string;
-  whileElementsMounted?: (
-    reference: RT,
-    floating: HTMLElement,
-    update: () => void
-  ) => void | CleanupFn;
-}
-
-export function useFloating<RT extends ReferenceType = ReferenceType>({
-  open = false,
-  onOpenChange = () => {},
-  whileElementsMounted,
-  placement,
-  middleware,
-  strategy,
-  nodeId,
-}: Partial<Props> = {}): UseFloatingReturn<RT> & {
-  context: FloatingContext<RT>;
-} {
-  const tree = useFloatingTree<RT>();
-  const dataRef = React.useRef<ContextData>({});
-  const events = React.useState(() => createPubSub())[0];
-  const floating = usePositionalFloating<RT>({
-    placement,
-    middleware,
-    strategy,
-    whileElementsMounted,
-  });
-
-  const context = React.useMemo<FloatingContext<RT>>(
-    () => ({
-      ...floating,
-      dataRef,
-      nodeId,
-      events,
-      open,
-      onOpenChange,
-    }),
-    [floating, dataRef, nodeId, events, open, onOpenChange]
-  );
-
-  useLayoutEffect(() => {
-    const node = tree?.nodesRef.current.find((node) => node.id === nodeId);
-    if (node) {
-      node.context = context;
-    }
-  });
-
-  return React.useMemo(
-    () => ({
-      context,
-      ...floating,
-    }),
-    [floating, context]
-  );
-}
-
 export * from '@floating-ui/react-dom';
+
+export {useFloating} from './useFloating';
 export {useInteractions} from './useInteractions';
 export {safePolygon} from './safePolygon';
 export {FloatingPortal, useFloatingPortalNode} from './FloatingPortal';
