+++ packages/react-dom/src/utils/useLatestRef.ts
@@ -1,0 +1,10 @@
+import * as React from 'react';
+import useLayoutEffect from 'use-isomorphic-layout-effect';
+
+export function useLatestRef<T>(value: T) {
+  const ref = React.useRef(value);
+  useLayoutEffect(() => {
+    ref.current = value;
+  });
+  return ref;
+}
